pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.23",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-stubs[compatible-mypy]>=5.1.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.7",
    "pillow>=11.3.0",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "backend.settings"